rsexpr

Enum OwnedSexpr

pub enum OwnedSexpr {
    List(OwnedSexprs),
    Group(OwnedSexprs),
    String(Vec<u8>),
    Atom(Vec<u8>),
    Comment(Vec<u8>),
}

An owned version of Sexpr. You can convert to and from Sexpr using the From trait.

Display

OwnedSexpr implements the Display trait for serializing to strings. By default, the output will try to minimize the amount of spaces used and the resulting output will be on one line. Enabling the formatterโs alternate flag using #, causes the output to be human-friendly / pretty-printed. Setting the precision with . additionally allows to specify the number of spaces used for indentation (2 by default).

For example:

let sexpr = rsexpr::from_slice(b"[ a b c ]").unwrap();
assert_eq!(format!("{sexpr}"), "[a b c]");
assert_eq!(format!("{sexpr:#}"), "[
  a
  b
  c
]");
assert_eq!(format!("{sexpr:#.4}"), "[
    a
    b
    c
]");

Variants

List(OwnedSexprs)

A list of OwnedSexprs surrounded by parentheses (, )

Group(OwnedSexprs)

A list of OwnedSexprs surrounded by brackets [, ]

String(Vec<u8>)

A sequence of bytes surrounded by quotes "

Atom(Vec<u8>)

A sequence of bytes not including whitespace, parens, and quotes

Comment(Vec<u8>)

Available on crate feature comments only.

A line comment, including the leading ;

Implementations

impl OwnedSexpr

pub fn unwrap_list(self) -> OwnedSexprs

Returns the contained List value, consuming self.

Panics

Panics if self is not List.

pub fn unwrap_list_ref(&self) -> &OwnedSexprs

Returns the contained List value by reference.

Panics

Panics if self is not List.

pub fn unwrap_group(self) -> OwnedSexprs

Returns the contained Group value, consuming self.

Panics

Panics if self is not Group.

pub fn unwrap_group_ref(&self) -> &OwnedSexprs

Returns the contained Group value by reference.

Panics

Panics if self is not Group.

pub fn unwrap_string(self) -> Vec<u8>

Returns the contained String value, consuming self.

Panics

Panics if self is not String.

pub fn unwrap_string_ref(&self) -> &Vec<u8>

Returns the contained String value by reference.

Panics

Panics if self is not String.

pub fn unwrap_atom(self) -> Vec<u8>

Returns the contained Atom value, consuming self.

Panics

Panics if self is not Atom.

pub fn unwrap_atom_ref(&self) -> &Vec<u8>

Returns the contained Atom value by reference.

Panics

Panics if self is not Atom.