Enum rsexpr::Sexpr

pub enum Sexpr<'src> {
    List(Sexprs<'src>),
    Group(Sexprs<'src>),
    String(Cow<'src, [u8]>),
    Atom(&'src [u8]),
    Comment(&'src [u8]),
}

A single node of the tree. The Atom and String variants reference the input slice. For an owned version have a look at OwnedSexpr.

Display

Sexpr implements the Display trait for serializing to strings. By default, the output will try to minimize the amount of spaces used and the resulting output will be on one line. Enabling the formatter’s alternate flag using #, causes the output to be human-friendly / pretty-printed. Setting the precision with . additionally allows to specify the number of spaces used for indentation (2 by default).

For example:

let sexpr = rsexpr::from_slice(b"[ a b c ]").unwrap();
assert_eq!(format!("{sexpr}"), "[a b c]");
assert_eq!(format!("{sexpr:#}"), "[
  a
  b
  c
]");
assert_eq!(format!("{sexpr:#.4}"), "[
    a
    b
    c
]");

Variants

List(Sexprs<'src>)

A list of Sexprs surrounded by parentheses (, )

Group(Sexprs<'src>)

A list of Sexprs surrounded by brackets [, ]

String(Cow<'src, [u8]>)

A sequence of bytes surrounded by quotes "

Atom(&'src [u8])

A sequence of bytes not including whitespace, parens, and quotes

Comment(&'src [u8])

Available on crate feature comments only.

A line comment, including the leading ;

Implementations

impl<'src> Sexpr<'src>

pub fn unwrap_list(self) -> Sexprs<'src>

Returns the contained List value, consuming self.

Panics

Panics if self is not List.

pub fn unwrap_list_ref(&self) -> &Sexprs<'src>

Returns the contained List value by reference.

Panics

Panics if self is not List.

pub fn unwrap_group(self) -> Sexprs<'src>

Returns the contained Group value, consuming self.

Panics

Panics if self is not Group.

pub fn unwrap_group_ref(&self) -> &Sexprs<'src>

Returns the contained Group value by reference.

Panics

Panics if self is not Group.

pub fn unwrap_string(self) -> Cow<'src, [u8]>

Returns the contained String value, consuming self.

Panics

Panics if self is not String.

pub fn unwrap_string_ref(&self) -> &Cow<'src, [u8]>

Returns the contained String value by reference.

Panics

Panics if self is not String.

pub fn unwrap_atom(self) -> &'src [u8]

Returns the contained Atom value, consuming self.

Panics

Panics if self is not Atom.

pub fn unwrap_atom_ref(&self) -> &&'src [u8]

Returns the contained Atom value by reference.

Panics

Panics if self is not Atom.